requirements.txt
requests
python-dotenv
PyYAML
pydantic
SQLAlchemy
PyMySQL
tenacity
pandas
pytest
cryptography
